Update launch-linux to check for NixOS
This change makes the shebang shell agnostic and adds logic to check for a NixOS system. If a NixOS system is present, check for Java in the nix store as declared by the NixOS system/flake, if present, utilize that instead of the dynamic java environment provided. If Java 21 is not present or of a mismatched version, output basic instructions for resolution. If not using NixOS stick to the original installation process for legacy linux distributions. Also sets up proper library paths for NixOS and enable optimizations.