Update download-java for NixOS
This change makes the shebang shell agnostic and adds logic to check if the system is running NixOS, if present skip the download as Java is being provided in the nix store due to declaration in the system/flake configuration. It also checks that the version of Java is Java 21, if not output basic instructions for resolution. If not running on NixOS follow the original installation of the dynamic Java environment.